栏目完成添加

+++ src/main/java/com/github/niefy/modules/cms/dao/CmsChannelMapper.java
@@ -13,5 +13,5 @@ import org.apache.ibatis.annotations.Mapper;
  */
 @Mapper
 public interface CmsChannelMapper extends BaseMapper<CmsChannelEntity> {
-    CmsChannelEntity loadChannelDetail(Integer channelId);
+    CmsChannelEntity loadChannelDetail(Long channelId);
 }
